Update model identifiers with aditional models.
